+++ data_prep/R/neighborhoods_2022.R
@@ -3,6 +3,7 @@ library(data.table)
 library(dplyr)
 library(furrr)
 
+
 year <- 2022
 
 # create dest dir