+++ r-package/NEWS.md
@@ -1,3 +1,12 @@
+# geobr dev
+
+**New data**
+
+
+- neighborhoods 2022
+- updated census tracts for 2022
+
+
 # geobr v1.9.1